Folders structure creation.
