+++ pages/home.py
@@ -1,0 +1,4 @@
+import streamlit as st
+
+def show_home():
+    st.title("Home")
