+++ pages/moreGameInfo.py
@@ -1,0 +1,4 @@
+import streamlit as st
+
+def show_more_game_info():
+    st.title("3333")
