+++ requirements.txt
@@ -7,7 +7,6 @@
 streamlit
 requests
 
-
 # If you want to display datasets, or if your API returns you dataframes,
 # you might need to add some extra stuff hereunder, e.g. pandas
